terminal: fix 32 bit build
I was trying to use vctl which uses this library on a 32 bit machine
(Linux 3.14.0 armv7l) and failed to build.

    ../../buger/goterm/terminal.go:81: constant 2147483648 overflows int
    ../../buger/goterm/terminal.go:85: constant 2147483648 overflows int

This will cut the constant in half on 32 bit systems. I don't know if
this is right but it seemed to work. Happy to fix up the patch further,
I just don't understand what PCT is for.