Add 24bit color support (#43)
* "RESET_COLOR" is default fg and bg color

According to the standard, colors 39 and 49 can be used to switch
to the terminal default colors for foreground and background
respectively. RESET_COLOR now uses 39;49 instead of 32, which
only sets the foreground color to green.

* Add 24bit color support

Added support for 24bit foreground and background colors with
CSI[38;2;<r>;<g>;<b>m and CSI[48;2;<r>;<g>;<b>m through the
GetRGBColor() and GetBgRGBColor() functions.

* Add functions to add 24bit color to strings

Oh holy crap I didn't realize getColor and getBgColor weren't
exported. Seems a little weird but here I'll add ColorRGB and
BackgroundRGB to format strings in 24bit color.

* Export GetColor functions

You know what how about we just export these, is there any reason
to enforce wrapping it around a string line by line? I can't think
of any reason to not allow the user to handle it.

* Make GetRGBColor functions more consistent

Since the ColorRGB and BackgroundRGB functions have the RGB as a
suffix we should probably do the same for the GetRGBColor functions
as well.

* Complete public GetColor transition

Missed a spot.